fix(core): critical bug fixes — test, SOS, api_retry, billing, security
- Fix test: use correct property name prescriptionRequired (was requiresPrescription)
- Fix SOS: use VHHttpClient.post() with proper JSON encoding + Content-Type
- Fix api_retry: cast pow() result to int for Duration multiplication
- Fix BillingInvoice: safe double parsing handles int, double, and String
- Add SecurityConfig.verifyOrWarn() — throws in production if cert pinning
  enabled but fingerprints list is empty

+++ lib/services/api_retry.dart
@@ -21,7 +21,7 @@ class ApiRetry {
         if (attempt >= maxRetries) rethrow;
         if (shouldRetry != null && !shouldRetry(e)) rethrow;
 
-        final delay = initialDelay * pow(2, attempt - 1);
+        final delay = initialDelay * pow(2, attempt - 1).toInt();
         if (kDebugMode) {
           debugPrint('ApiRetry: attempt $attempt failed, retrying in ${delay.inMilliseconds}ms');
         }
